+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix the percentage overflow in kernel module table
+- Use the default colors if the accent color is not provided
 
 ## [1.4.0] - 2020-08-05
 ### Added